+++ src/components/Project/Resource/DFR3MappingPage.tsx
@@ -151,10 +151,10 @@ const DFR3MappingPage = (): JSX.Element => {
                         <ProjectHeader project={project} />
                         <Divider />
                         <Grid container spacing={5} mt={3} ml={0}>
-                            <Grid sm={2}>
+                            <Grid sm={3}>
                                 <ProjectSidebar id={project.id} />
                             </Grid>
-                            <Grid sm={10}>
+                            <Grid sm={9}>
                                 <ResourceFilterBar
                                     title="DFR3Mappings"
                                     icon={<DFR3Icon sx={{ verticalAlign: "middle" }} />}